requirements.txt
streamlit
langchain-community
pdfplumber
python-docx
spacy
python-dotenv
langchain
langchain-core
langchain-groq
